add tag
You'll be tagging the SaveImageToFileWorker WorkRequest, so that you can get it using getWorkInfosByTag. You'll use a tag to label your work instead of using the WorkManager ID, because if your user blurs multiple images, all of the saving image WorkRequests will have the same tag but not the same ID. Also you are able to pick the tag.